Refactored the code according to feedback
