+++ src/apps/backend/app.ts
@@ -11,7 +11,7 @@ import { ConfigService } from './modules/config';
 import { Logger, CustomLoggerTransport } from './modules/logger';
 import { PasswordResetTokenServer } from './modules/password-reset-token';
 import { TaskServer } from './modules/task';
-import { CommentServer } from './modules/comment';
+import { CommentServer } from './modules/task-comment';
 
 const isDevEnv = process.env.NODE_ENV === 'development';
