--- src/apps/backend/modules/comment/index.ts
@@ -1,4 +1,0 @@
-export { default as CommentServer } from './rest-api/comment-server';
-
-export { default as CommentService } from './comment-service';
-export * from './types';
