+++ src/apps/backend/modules/task-comment/index.ts
@@ -1,0 +1,4 @@
+export { default as CommentServer } from './rest-api/task-comment-server';
+
+export { default as CommentService } from './task-comment-service';
+export * from './types';
