+++ src/apps/backend/modules/task-comment/internal/store/task-comment-repository.ts
@@ -1,6 +1,6 @@
 import { ApplicationRepository } from '../../../application';
 
-import { CommentDB, CommentDbSchema } from './comment-db';
+import { CommentDB, CommentDbSchema } from './task-comment-db';
 
 const CommentRepository = ApplicationRepository<CommentDB>(
   'Comment',
